Extract BlockAnalyzer from SteepBridge
